+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/chunks/Cm4ryvZL.js
@@ -1,1 +1,0 @@
-import"./-6hovnNO.js";import{i as e}from"./fra9WeKx.js";var t=()=>{let t=e;return{page:{subscribe:t.page.subscribe},navigating:{subscribe:t.navigating.subscribe},updated:t.updated}},n={subscribe(e){return t().page.subscribe(e)}};export{n as t};