+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/chunks/Cv3bqXI0.js
@@ -1,0 +1,1 @@
+import"./-6hovnNO.js";import{i as e}from"./3aiihaAk.js";var t=()=>{let t=e;return{page:{subscribe:t.page.subscribe},navigating:{subscribe:t.navigating.subscribe},updated:t.updated}},n={subscribe(e){return t().page.subscribe(e)}};export{n as t};