+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/entry/start.CLkZsENJ.js
@@ -1,1 +1,0 @@
-import{r as e,s as t}from"../chunks/fra9WeKx.js";export{t as load_css,e as start};